requirements.txt
Brotli==1.0.9
click==7.1.2
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
matlabengineforpython===R2019b
numpy==1.19.5
pandas==1.2.0
plotly==4.14.3
python-dateutil==2.8.1
pytz==2020.5
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
